introduced the concept of the VersionedAggregate a convenience class to reduce the complexity of managing versioning in Aggregates.

+++ src/main/java/com/eventjournal/api/Aggregate.java
@@ -7,5 +7,5 @@ public interface Aggregate {
         return StreamId.of(getClass(), getId());
     }
 
-    default Integer version() { return 0; };
+    default long version() { return 0L; };
 }
