+++ src/main/java/com/eventjournal/api/Header.java
@@ -10,10 +10,31 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 /**
- * The Header contains metadata for a Message.
+ * The Header contains metadata for a Message, and are required when constructing a Message.
  * Use the Header to store your own additional metadata
- * for a message if need by with the put(key, value) method.
+ * for a message if needed by way of the put(key, value) method.
  * Later when you need to retrieve the value, use the get(key) method.
+ * The Header offers several opinionated static constructors for
+ * creating a new Header for a Message. The Header is immutable.
+ * <p>
+ * Opinions and concepts:
+ * HeadOfChain - the first (and possibly only) message in a chain of messages
+ * ResultingFrom - a message that is a direct result of another message - Used to indicate an Event is the result of a Command
+ * SideEffectOf - a message that is a side effect of another message - Used to indicate a Command is a Side Effect of an Event
+ * <p>
+ * These static constructors ensure that developers do not have to focus on tying correlation IDs from one message to another.
+ * The Header will automatically populate the correlation ID when using the static constructors.
+ *
+ * <p>
+ * example:
+ * <pre>
+ *         Header header = Header.headOfChain(myAggregate, MyEvent.class);
+ *         header.put("my-custom-header", "some-value");
+ *         header.get("my-custom-header").ifPresent(System.out::println);
+ *         // prints "some-value"
+ *         header.get("non-existent-key");
+ *         // returns Optional.empty()
+ * </pre>
  */
 public class Header {
     @JsonIgnore
@@ -22,7 +43,7 @@ public class Header {
     String messageType;
     Message.MessageCategory category;
     Instant timestamp;
-    Integer sequence;
+    long sequence;
     String causationId;
     String messageId;
     String correlationId;
@@ -36,14 +57,14 @@ public class Header {
                   String messageType,
                   Message.MessageCategory category,
                   Instant timestamp,
-                  Integer sequence,
+                  long sequence,
                   String causationId,
                   String messageId,
                   String correlationId) {
         establishProps(streamId, messageType, category, timestamp, sequence, causationId, messageId, correlationId);
     }
 
-    private void establishProps(String streamId, String messageType, Message.MessageCategory category, Instant timestamp, Integer version, String causationId, String messageId, String messageChainId) {
+    private void establishProps(String streamId, String messageType, Message.MessageCategory category, Instant timestamp, long version, String causationId, String messageId, String messageChainId) {
         this.streamId = Objects.requireNonNull(streamId, "Stream ID is required");
         this.messageType = Objects.requireNonNull(messageType, "Message Type is required");
         this.timestamp = Objects.requireNonNull(timestamp, "Timestamp is required!");
@@ -130,7 +151,7 @@ public class Header {
      * @return a new MessageHeader
      */
     public static Header headOfChain(Aggregate aggregate, Class<? extends Message> messageType) {
-        return new Header(StreamId.of(aggregate.getClass(), aggregate.getId()), messageType, aggregate.version() + 1);
+        return new Header(StreamId.of(aggregate.getClass(), aggregate.getId()), messageType, aggregate.version() + 1L);
     }
 
     /**
@@ -186,7 +207,7 @@ public class Header {
      * @param messageType the MessageType (some implementation of a Command or Event class.simpleName())
      * @param sequence    the aggregate version for which this message was raised
      */
-    public Header(String streamId, Class<? extends Message> messageType, int sequence) {
+    public Header(String streamId, Class<? extends Message> messageType, long sequence) {
         this.streamId = Objects.requireNonNull(streamId, "Stream ID is required");
         this.messageType = Objects.requireNonNull(messageType, "Message Type is required").getSimpleName();
         this.sequence = sequence;
@@ -268,7 +289,7 @@ public class Header {
         return Instant.ofEpochMilli(timestamp.toEpochMilli());
     }
 
-    public Integer getSequence() {
+    public long getSequence() {
         return sequence;
     }
